orbit: new API for validity, providing more precise test of trajectory type
is_valid() is a generalization of existing is_orbit() and new is_precise_placement().
is_valid() will be used by client code to determine if the trajectory is
able to spawn something.
The two others are used internally to determined which method should be used
to place the astre at their right place.